requirements.txt
androguard==4.1.3
PyYAML==6.0.2
